Add warning to the top of README.win32
Since .vcproj aren't generated yet, following the readme without using cmake to make them is confusing. This warning doesn't fix the problem, but it does steer users in the right direction.

The vc9 directories and so-on, continue to have solutions in them. So, it appears like they can be built. They probably also need deleting, unless cmake is relying on them as input.